Retry blobstore saves
Fog local isn't safe for concurrent use meaning that if a blob is deleted at
the same time that one is saved it is possible for the delete to remove the
parent directories of the saving blob, causing the save to fail. Since we're using
the blobstore from multiple processes we can't take an in-process lock, and we
don't want to get in to using file locking over NFS, so instead we're going to just
retry saving a few times which should avoid the issue.

Signed-off-by: Michael Fraenkel <fraenkel@us.ibm.com>